Crate google_tagmanager1

This documentation was generated from Tag Manager crate version 1.0.0+20160310, where 20160310 is the exact revision of the tagmanager:v1 schema built by the mako code generator v1.0.0.

Everything else about the Tag Manager v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

accounts

containers create, containers delete, containers environments create, containers environments delete, containers environments get, containers environments list, containers environments patch, containers environments update, containers folders create, containers folders delete, containers folders entities list, containers folders get, containers folders list, containers folders update, containers get, containers list, containers move_folders update, containers reauthorize_environments update, containers tags create, containers tags delete, containers tags get, containers tags list, containers tags update, containers triggers create, containers triggers delete, containers triggers get, containers triggers list, containers triggers update, containers update, containers variables create, containers variables delete, containers variables get, containers variables list, containers variables update, containers versions create, containers versions delete, containers versions get, containers versions list, containers versions publish, containers versions restore, containers versions undelete, containers versions update, get, list, permissions create, permissions delete, permissions get, permissions list, permissions update and update

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.accounts().containers_list(...).doit()
let r = hub.accounts().permissions_list(...).doit()
let r = hub.accounts().containers_environments_create(...).doit()
let r = hub.accounts().permissions_create(...).doit()
let r = hub.accounts().containers_environments_delete(...).doit()
let r = hub.accounts().permissions_delete(...).doit()
let r = hub.accounts().containers_get(...).doit()
let r = hub.accounts().containers_environments_patch(...).doit()
let r = hub.accounts().containers_versions_list(...).doit()
let r = hub.accounts().containers_triggers_update(...).doit()
let r = hub.accounts().containers_triggers_get(...).doit()
let r = hub.accounts().containers_delete(...).doit()
let r = hub.accounts().containers_folders_delete(...).doit()
let r = hub.accounts().containers_create(...).doit()
let r = hub.accounts().containers_tags_delete(...).doit()
let r = hub.accounts().containers_folders_entities_list(...).doit()
let r = hub.accounts().containers_versions_undelete(...).doit()
let r = hub.accounts().containers_environments_list(...).doit()
let r = hub.accounts().containers_tags_list(...).doit()
let r = hub.accounts().containers_versions_publish(...).doit()
let r = hub.accounts().containers_folders_get(...).doit()
let r = hub.accounts().containers_environments_get(...).doit()
let r = hub.accounts().containers_tags_create(...).doit()
let r = hub.accounts().containers_triggers_list(...).doit()
let r = hub.accounts().containers_versions_delete(...).doit()
let r = hub.accounts().update(...).doit()
let r = hub.accounts().containers_versions_create(...).doit()
let r = hub.accounts().permissions_get(...).doit()
let r = hub.accounts().containers_move_folders_update(...).doit()
let r = hub.accounts().containers_versions_restore(...).doit()
let r = hub.accounts().containers_variables_create(...).doit()
let r = hub.accounts().containers_variables_list(...).doit()
let r = hub.accounts().containers_tags_get(...).doit()
let r = hub.accounts().containers_variables_get(...).doit()
let r = hub.accounts().containers_triggers_delete(...).doit()
let r = hub.accounts().containers_triggers_create(...).doit()
let r = hub.accounts().containers_folders_update(...).doit()
let r = hub.accounts().list(...).doit()
let r = hub.accounts().permissions_update(...).doit()
let r = hub.accounts().containers_variables_delete(...).doit()
let r = hub.accounts().containers_reauthorize_environments_update(...).doit()
let r = hub.accounts().get(...).doit()
let r = hub.accounts().containers_update(...).doit()
let r = hub.accounts().containers_tags_update(...).doit()
let r = hub.accounts().containers_environments_update(...).doit()
let r = hub.accounts().containers_folders_create(...).doit()
let r = hub.accounts().containers_folders_list(...).doit()
let r = hub.accounts().containers_versions_update(...).doit()
let r = hub.accounts().containers_variables_update(...).doit()
let r = hub.accounts().containers_versions_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-tagmanager1 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_tagmanager1 as tagmanager1;
use tagmanager1::Folder;
use tagmanager1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use tagmanager1::TagManager;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = TagManager::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Folder::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().containers_move_folders_update(req, "accountId", "containerId", "folderId")
             .add_variable_id("clita")
             .add_trigger_id("invidunt")
             .add_tag_id("eirmod")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Account
Represents a Google Tag Manager Account.

AccountAccess
Defines the Google Tag Manager Account access permissions.

AccountContainerCreateCall
Creates a Container.

AccountContainerDeleteCall
Deletes a Container.

AccountContainerEnvironmentCreateCall
Creates a GTM Environment.

AccountContainerEnvironmentDeleteCall
Deletes a GTM Environment.

AccountContainerEnvironmentGetCall
Gets a GTM Environment.

AccountContainerEnvironmentListCall
Lists all GTM Environments of a GTM Container.

AccountContainerEnvironmentPatchCall
Updates a GTM Environment. This method supports patch semantics.

AccountContainerEnvironmentUpdateCall
Updates a GTM Environment.

AccountContainerFolderCreateCall
Creates a GTM Folder.

AccountContainerFolderDeleteCall
Deletes a GTM Folder.

AccountContainerFolderEntityListCall
List all entities in a GTM Folder.

AccountContainerFolderGetCall
Gets a GTM Folder.

AccountContainerFolderListCall
Lists all GTM Folders of a Container.

AccountContainerFolderUpdateCall
Updates a GTM Folder.

AccountContainerGetCall
Gets a Container.

AccountContainerListCall
Lists all Containers that belongs to a GTM Account.

AccountContainerMoveFolderUpdateCall
Moves entities to a GTM Folder.

AccountContainerReauthorizeEnvironmentUpdateCall
Re-generates the authorization code for a GTM Environment.

AccountContainerTagCreateCall
Creates a GTM Tag.

AccountContainerTagDeleteCall
Deletes a GTM Tag.

AccountContainerTagGetCall
Gets a GTM Tag.

AccountContainerTagListCall
Lists all GTM Tags of a Container.

AccountContainerTagUpdateCall
Updates a GTM Tag.

AccountContainerTriggerCreateCall
Creates a GTM Trigger.

AccountContainerTriggerDeleteCall
Deletes a GTM Trigger.

AccountContainerTriggerGetCall
Gets a GTM Trigger.

AccountContainerTriggerListCall
Lists all GTM Triggers of a Container.

AccountContainerTriggerUpdateCall
Updates a GTM Trigger.

AccountContainerUpdateCall
Updates a Container.

AccountContainerVariableCreateCall
Creates a GTM Variable.

AccountContainerVariableDeleteCall
Deletes a GTM Variable.

AccountContainerVariableGetCall
Gets a GTM Variable.

AccountContainerVariableListCall
Lists all GTM Variables of a Container.

AccountContainerVariableUpdateCall
Updates a GTM Variable.

AccountContainerVersionCreateCall
Creates a Container Version.

AccountContainerVersionDeleteCall
Deletes a Container Version.

AccountContainerVersionGetCall
Gets a Container Version.

AccountContainerVersionListCall
Lists all Container Versions of a GTM Container.

AccountContainerVersionPublishCall
Publishes a Container Version.

AccountContainerVersionRestoreCall
Restores a Container Version. This will overwrite the container's current configuration (including its variables, triggers and tags). The operation will not have any effect on the version that is being served (i.e. the published version).

AccountContainerVersionUndeleteCall
Undeletes a Container Version.

AccountContainerVersionUpdateCall
Updates a Container Version.

AccountGetCall
Gets a GTM Account.

AccountListCall
Lists all GTM Accounts that a user has access to.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the TagManager hub.

AccountPermissionCreateCall
Creates a user's Account & Container Permissions.

AccountPermissionDeleteCall
Removes a user from the account, revoking access to it and all of its containers.

AccountPermissionGetCall
Gets a user's Account & Container Permissions.

AccountPermissionListCall
List all users that have access to the account along with Account and Container Permissions granted to each of them.

AccountPermissionUpdateCall
Updates a user's Account & Container Permissions.

AccountUpdateCall
Updates a GTM Account.

Condition
Represents a predicate.

Container
Represents a Google Tag Manager Container.

ContainerAccess
Defines the Google Tag Manager Container access permissions.

ContainerVersion
Represents a Google Tag Manager Container Version.

ContainerVersionHeader
Represents a Google Tag Manager Container Version Header.

CreateContainerVersionRequestVersionOptions
Options for new container versions.

CreateContainerVersionResponse
Create container versions response.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Environment
Represents a Google Tag Manager Environment. Note that a user can create, delete and update environments of type USER, but can only update the enable_debug and url fields of environments of other types.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

Folder
Represents a Google Tag Manager Folder.

FolderEntities
Represents a Google Tag Manager Folder's contents.

ListAccountUsersResponse
List AccountUsers Response.

ListAccountsResponse
List Accounts Response.

ListContainerVersionsResponse
List container versions response.

ListContainersResponse
List Containers Response.

ListEnvironmentsResponse
List Environments Response.

ListFoldersResponse
List Folders Response.

ListTagsResponse
List Tags Response.

ListTriggersResponse
List triggers response.

ListVariablesResponse
List Variables Response.

Macro
Represents a Google Tag Manager Macro.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Parameter
Represents a Google Tag Manager Parameter.

PublishContainerVersionResponse
Publish container version response.

Rule
Represents a Google Tag Manager Rule.

SetupTag
There is no detailed description.

Tag
Represents a Google Tag Manager Tag.

TagManager
Central instance to access all TagManager related resource activities

TeardownTag
There is no detailed description.

Trigger
Represents a Google Tag Manager Trigger

UserAccess
Represents a user's permissions to an account and its container.

Variable
Represents a Google Tag Manager Variable.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.