Module google_tagmanager1::api

Structs

Account
Represents a Google Tag Manager Account.

AccountAccess
Defines the Google Tag Manager Account access permissions.

AccountContainerCreateCall
Creates a Container.

AccountContainerDeleteCall
Deletes a Container.

AccountContainerEnvironmentCreateCall
Creates a GTM Environment.

AccountContainerEnvironmentDeleteCall
Deletes a GTM Environment.

AccountContainerEnvironmentGetCall
Gets a GTM Environment.

AccountContainerEnvironmentListCall
Lists all GTM Environments of a GTM Container.

AccountContainerEnvironmentUpdateCall
Updates a GTM Environment.

AccountContainerFolderCreateCall
Creates a GTM Folder.

AccountContainerFolderDeleteCall
Deletes a GTM Folder.

AccountContainerFolderEntityListCall
List all entities in a GTM Folder.

AccountContainerFolderGetCall
Gets a GTM Folder.

AccountContainerFolderListCall
Lists all GTM Folders of a Container.

AccountContainerFolderUpdateCall
Updates a GTM Folder.

AccountContainerGetCall
Gets a Container.

AccountContainerListCall
Lists all Containers that belongs to a GTM Account.

AccountContainerMoveFolderUpdateCall
Moves entities to a GTM Folder.

AccountContainerReauthorizeEnvironmentUpdateCall
Re-generates the authorization code for a GTM Environment.

AccountContainerTagCreateCall
Creates a GTM Tag.

AccountContainerTagDeleteCall
Deletes a GTM Tag.

AccountContainerTagGetCall
Gets a GTM Tag.

AccountContainerTagListCall
Lists all GTM Tags of a Container.

AccountContainerTagUpdateCall
Updates a GTM Tag.

AccountContainerTriggerCreateCall
Creates a GTM Trigger.

AccountContainerTriggerDeleteCall
Deletes a GTM Trigger.

AccountContainerTriggerGetCall
Gets a GTM Trigger.

AccountContainerTriggerListCall
Lists all GTM Triggers of a Container.

AccountContainerTriggerUpdateCall
Updates a GTM Trigger.

AccountContainerUpdateCall
Updates a Container.

AccountContainerVariableCreateCall
Creates a GTM Variable.

AccountContainerVariableDeleteCall
Deletes a GTM Variable.

AccountContainerVariableGetCall
Gets a GTM Variable.

AccountContainerVariableListCall
Lists all GTM Variables of a Container.

AccountContainerVariableUpdateCall
Updates a GTM Variable.

AccountContainerVersionCreateCall
Creates a Container Version.

AccountContainerVersionDeleteCall
Deletes a Container Version.

AccountContainerVersionGetCall
Gets a Container Version.

AccountContainerVersionListCall
Lists all Container Versions of a GTM Container.

AccountContainerVersionPublishCall
Publishes a Container Version.

AccountContainerVersionRestoreCall
Restores a Container Version. This will overwrite the container's current configuration (including its variables, triggers and tags). The operation will not have any effect on the version that is being served (i.e. the published version).

AccountContainerVersionUndeleteCall
Undeletes a Container Version.

AccountContainerVersionUpdateCall
Updates a Container Version.

AccountGetCall
Gets a GTM Account.

AccountListCall
Lists all GTM Accounts that a user has access to.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the TagManager hub.

AccountPermissionCreateCall
Creates a user's Account & Container Permissions.

AccountPermissionDeleteCall
Removes a user from the account, revoking access to it and all of its containers.

AccountPermissionGetCall
Gets a user's Account & Container Permissions.

AccountPermissionListCall
List all users that have access to the account along with Account and Container Permissions granted to each of them.

AccountPermissionUpdateCall
Updates a user's Account & Container Permissions.

AccountUpdateCall
Updates a GTM Account.

Condition
Represents a predicate.

Container
Represents a Google Tag Manager Container.

ContainerAccess
Defines the Google Tag Manager Container access permissions.

ContainerVersion
Represents a Google Tag Manager Container Version.

ContainerVersionHeader
Represents a Google Tag Manager Container Version Header.

CreateContainerVersionRequestVersionOptions
Options for new container versions.

CreateContainerVersionResponse
Create container versions response.

Environment
Represents a Google Tag Manager Environment. Note that a user can create, delete and update environments of type USER, but can only update the enable_debug and url fields of environments of other types.

Folder
Represents a Google Tag Manager Folder.

FolderEntities
Represents a Google Tag Manager Folder's contents.

ListAccountUsersResponse
List AccountUsers Response.

ListAccountsResponse
List Accounts Response.

ListContainerVersionsResponse
List container versions response.

ListContainersResponse
List Containers Response.

ListEnvironmentsResponse
List Environments Response.

ListFoldersResponse
List Folders Response.

ListTagsResponse
List Tags Response.

ListTriggersResponse
List triggers response.

ListVariablesResponse
List Variables Response.

Macro
Represents a Google Tag Manager Macro.

Parameter
Represents a Google Tag Manager Parameter.

PublishContainerVersionResponse
Publish container version response.

Rule
Represents a Google Tag Manager Rule.

SetupTag
There is no detailed description.

Tag
Represents a Google Tag Manager Tag.

TagManager
Central instance to access all TagManager related resource activities

TeardownTag
There is no detailed description.

Trigger
Represents a Google Tag Manager Trigger

UserAccess
Represents a user's permissions to an account and its container.

Variable
Represents a Google Tag Manager Variable.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.