Enum google_tagmanager1::Scope

pub enum Scope {
    ManageUser,
    ManageAccount,
    Readonly,
    EditContainerversion,
    EditContainer,
    DeleteContainer,
    Publish,
}

Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Variants
ManageUser

Manage user permissions of your Google Tag Manager data

ManageAccount

Manage your Google Tag Manager accounts

Readonly

View your Google Tag Manager containers

EditContainerversion

Manage your Google Tag Manager container versions

EditContainer

Manage your Google Tag Manager containers

DeleteContainer

Delete your Google Tag Manager containers

Publish

Publish your Google Tag Manager containers