Struct google-tagmanager1::AccountMethodsBuilder

pub struct AccountMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on account resources. It is not used directly, but through the TagManager hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-tagmanager1" as tagmanager1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use tagmanager1::TagManager;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = TagManager::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `containers_create(...)`, `containers_delete(...)`, `containers_get(...)`, `containers_list(...)`, `containers_macros_create(...)`, `containers_macros_delete(...)`, `containers_macros_get(...)`, `containers_macros_list(...)`, `containers_macros_update(...)`, `containers_rules_create(...)`, `containers_rules_delete(...)`, `containers_rules_get(...)`, `containers_rules_list(...)`, `containers_rules_update(...)`, `containers_tags_create(...)`, `containers_tags_delete(...)`, `containers_tags_get(...)`, `containers_tags_list(...)`, `containers_tags_update(...)`, `containers_triggers_create(...)`, `containers_triggers_delete(...)`, `containers_triggers_get(...)`, `containers_triggers_list(...)`, `containers_triggers_update(...)`, `containers_update(...)`, `containers_variables_create(...)`, `containers_variables_delete(...)`, `containers_variables_get(...)`, `containers_variables_list(...)`, `containers_variables_update(...)`, `containers_versions_create(...)`, `containers_versions_delete(...)`, `containers_versions_get(...)`, `containers_versions_list(...)`, `containers_versions_publish(...)`, `containers_versions_restore(...)`, `containers_versions_undelete(...)`, `containers_versions_update(...)`, `get(...)`, `list(...)`, `permissions_create(...)`, `permissions_delete(...)`, `permissions_get(...)`, `permissions_list(...)`, `permissions_update(...)` and `update(...)`
// to build up your call.
let rb = hub.accounts();

Methods

impl<'a, C, NC, A> AccountMethodsBuilder<'a, C, NC, A>

fn containers_list(&self, account_id: &str) -> AccountContainerListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all Containers that belongs to a GTM Account.

fn permissions_list(&self, account_id: &str) -> AccountPermissionListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List all users that have access to the account along with Account and Container Permissions granted to each of them.

fn containers_versions_undelete(&self, account_id: &str, container_id: &str, container_version_id: &str) -> AccountContainerVersionUndeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Undeletes a Container Version.

fn permissions_create(&self, request: &UserAccess, account_id: &str) -> AccountPermissionCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a user's Account & Container Permissions.

fn permissions_delete(&self, account_id: &str, permission_id: &str) -> AccountPermissionDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes a user from the account, revoking access to it and all of its containers.

fn containers_get(&self, account_id: &str, container_id: &str) -> AccountContainerGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a Container.

fn containers_versions_list(&self, account_id: &str, container_id: &str) -> AccountContainerVersionListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all Container Versions of a GTM Container.

fn containers_triggers_update(&self, request: &Trigger, account_id: &str, container_id: &str, trigger_id: &str) -> AccountContainerTriggerUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a GTM Trigger.

fn containers_triggers_get(&self, account_id: &str, container_id: &str, trigger_id: &str) -> AccountContainerTriggerGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a GTM Trigger.

fn containers_delete(&self, account_id: &str, container_id: &str) -> AccountContainerDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a Container.

fn containers_create(&self, request: &Container, account_id: &str) -> AccountContainerCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a Container.

fn containers_tags_delete(&self, account_id: &str, container_id: &str, tag_id: &str) -> AccountContainerTagDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a GTM Tag.

fn containers_rules_update(&self, request: &Rule, account_id: &str, container_id: &str, rule_id: &str) -> AccountContainerRuleUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a GTM Rule.

fn containers_rules_delete(&self, account_id: &str, container_id: &str, rule_id: &str) -> AccountContainerRuleDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a GTM Rule.

fn containers_tags_list(&self, account_id: &str, container_id: &str) -> AccountContainerTagListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all GTM Tags of a Container.

fn containers_versions_publish(&self, account_id: &str, container_id: &str, container_version_id: &str) -> AccountContainerVersionPublishMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Publishes a Container Version.

fn containers_tags_create(&self, request: &Tag, account_id: &str, container_id: &str) -> AccountContainerTagCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a GTM Tag.

fn containers_triggers_list(&self, account_id: &str, container_id: &str) -> AccountContainerTriggerListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all GTM Triggers of a Container.

fn containers_versions_delete(&self, account_id: &str, container_id: &str, container_version_id: &str) -> AccountContainerVersionDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a Container Version.

fn update(&self, request: &Account, account_id: &str) -> AccountUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a GTM Account.

fn containers_macros_delete(&self, account_id: &str, container_id: &str, macro_id: &str) -> AccountContainerMacroDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a GTM Macro.

fn containers_versions_create(&self, request: &CreateContainerVersionRequestVersionOptions, account_id: &str, container_id: &str) -> AccountContainerVersionCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a Container Version.

fn permissions_get(&self, account_id: &str, permission_id: &str) -> AccountPermissionGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a user's Account & Container Permissions.

fn containers_rules_create(&self, request: &Rule, account_id: &str, container_id: &str) -> AccountContainerRuleCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a GTM Rule.

fn containers_versions_restore(&self, account_id: &str, container_id: &str, container_version_id: &str) -> AccountContainerVersionRestoreMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Restores a Container Version. This will overwrite the container's current configuration (including its macros, rules and tags). The operation will not have any effect on the version that is being served (i.e. the published version).

fn containers_rules_get(&self, account_id: &str, container_id: &str, rule_id: &str) -> AccountContainerRuleGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a GTM Rule.

fn containers_variables_create(&self, request: &Variable, account_id: &str, container_id: &str) -> AccountContainerVariableCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a GTM Variable.

fn containers_variables_list(&self, account_id: &str, container_id: &str) -> AccountContainerVariableListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all GTM Variables of a Container.

fn containers_macros_create(&self, request: &Macro, account_id: &str, container_id: &str) -> AccountContainerMacroCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a GTM Macro.

fn containers_tags_get(&self, account_id: &str, container_id: &str, tag_id: &str) -> AccountContainerTagGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a GTM Tag.

fn containers_variables_get(&self, account_id: &str, container_id: &str, variable_id: &str) -> AccountContainerVariableGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a GTM Variable.

fn containers_triggers_delete(&self, account_id: &str, container_id: &str, trigger_id: &str) -> AccountContainerTriggerDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a GTM Trigger.

fn containers_macros_list(&self, account_id: &str, container_id: &str) -> AccountContainerMacroListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all GTM Macros of a Container.

fn containers_triggers_create(&self, request: &Trigger, account_id: &str, container_id: &str) -> AccountContainerTriggerCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a GTM Trigger.

fn containers_macros_update(&self, request: &Macro, account_id: &str, container_id: &str, macro_id: &str) -> AccountContainerMacroUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a GTM Macro.

fn list(&self) -> AccountListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all GTM Accounts that a user has access to.

fn permissions_update(&self, request: &UserAccess, account_id: &str, permission_id: &str) -> AccountPermissionUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a user's Account & Container Permissions.

fn containers_variables_delete(&self, account_id: &str, container_id: &str, variable_id: &str) -> AccountContainerVariableDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a GTM Variable.

fn get(&self, account_id: &str) -> AccountGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a GTM Account.

fn containers_update(&self, request: &Container, account_id: &str, container_id: &str) -> AccountContainerUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a Container.

fn containers_rules_list(&self, account_id: &str, container_id: &str) -> AccountContainerRuleListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all GTM Rules of a Container.

fn containers_tags_update(&self, request: &Tag, account_id: &str, container_id: &str, tag_id: &str) -> AccountContainerTagUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a GTM Tag.

fn containers_macros_get(&self, account_id: &str, container_id: &str, macro_id: &str) -> AccountContainerMacroGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a GTM Macro.

fn containers_versions_update(&self, request: &ContainerVersion, account_id: &str, container_id: &str, container_version_id: &str) -> AccountContainerVersionUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a Container Version.

fn containers_variables_update(&self, request: &Variable, account_id: &str, container_id: &str, variable_id: &str) -> AccountContainerVariableUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a GTM Variable.

fn containers_versions_get(&self, account_id: &str, container_id: &str, container_version_id: &str) -> AccountContainerVersionGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a Container Version.