Struct google_tagmanager1::ContainerVersion

pub struct ContainerVersion {
    pub container: Option<Container>,
    pub container_id: Option<String>,
    pub trigger: Option<Vec<Trigger>>,
    pub deleted: Option<bool>,
    pub macro_: Option<Vec<Macro>>,
    pub notes: Option<String>,
    pub rule: Option<Vec<Rule>>,
    pub tag: Option<Vec<Tag>>,
    pub container_version_id: Option<String>,
    pub fingerprint: Option<String>,
    pub variable: Option<Vec<Variable>>,
    pub folder: Option<Vec<Folder>>,
    pub account_id: Option<String>,
    pub name: Option<String>,
}

Represents a Google Tag Manager Container Version.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers versions update accounts (request|response)

containers versions get accounts (response)

containers versions undelete accounts (response)

containers versions restore accounts (response)

Fields

container

The container that this version was taken from.

container_id

GTM Container ID.

trigger

The triggers in the container that this version was taken from.

deleted

A value of true indicates this container version has been deleted.

macro_

The macros in the container that this version was taken from.

notes

User notes on how to apply this container version in the container.

rule

The rules in the container that this version was taken from.

tag

The tags in the container that this version was taken from.

container_version_id

The Container Version ID uniquely identifies the GTM Container Version.

fingerprint

The fingerprint of the GTM Container Version as computed at storage time. This value is recomputed whenever the container version is modified.

variable

The variables in the container that this version was taken from.

folder

The folders in the container that this version was taken from.

account_id

GTM Account ID.

name

Container version display name.