Struct google_tagmanager1::ContainerVersionHeader

pub struct ContainerVersionHeader {
    pub name: Option<String>,
    pub deleted: Option<bool>,
    pub num_macros: Option<String>,
    pub container_id: Option<String>,
    pub num_tags: Option<String>,
    pub num_rules: Option<String>,
    pub container_version_id: Option<String>,
    pub num_triggers: Option<String>,
    pub account_id: Option<String>,
    pub num_variables: Option<String>,
}

Represents a Google Tag Manager Container Version Header.

This type is not used in any activity, and only used as part of another schema.

Fields

name

Container version display name.

deleted

A value of true indicates this container version has been deleted.

num_macros

Number of macros in the container version.

container_id

GTM Container ID.

num_tags

Number of tags in the container version.

num_rules

Number of rules in the container version.

container_version_id

The Container Version ID uniquely identifies the GTM Container Version.

num_triggers

Number of triggers in the container version.

account_id

GTM Account ID.

num_variables

Number of variables in the container version.