Struct google_tagmanager1::Environment

pub struct Environment {
    pub authorization_code: Option<String>,
    pub description: Option<String>,
    pub environment_id: Option<String>,
    pub url: Option<String>,
    pub enable_debug: Option<bool>,
    pub container_id: Option<String>,
    pub authorization_timestamp_ms: Option<String>,
    pub container_version_id: Option<String>,
    pub fingerprint: Option<String>,
    pub account_id: Option<String>,
    pub type_: Option<String>,
    pub name: Option<String>,
}

Represents a Google Tag Manager Environment. Note that a user can create, delete and update environments of type USER, but can only update the enable_debug and url fields of environments of other types.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers environments create accounts (request|response)

containers environments update accounts (request|response)

containers environments patch accounts (request|response)

containers environments get accounts (response)

containers reauthorize_environments update accounts (request|response)

Fields
authorization_code: Option<String>

The environment authorization code.

description: Option<String>

The environment description. Can be set or changed only on USER type environments.

environment_id: Option<String>

GTM Environment ID uniquely identifies the GTM Environment.

url: Option<String>

Default preview page url for the environment.

enable_debug: Option<bool>

Whether or not to enable debug by default on for the environment.

container_id: Option<String>

GTM Container ID.

authorization_timestamp_ms: Option<String>

The last update time-stamp for the authorization code.

container_version_id: Option<String>

fingerprint: Option<String>

The fingerprint of the GTM environment as computed at storage time. This value is recomputed whenever the environment is modified.

account_id: Option<String>

GTM Account ID.

type_: Option<String>

The type of this environment.

name: Option<String>

The environment display name. Can be set or changed only on USER type environments.