Struct google_tagmanager1::Tag

pub struct Tag {
    pub schedule_start_ms: Option<String>,
    pub schedule_end_ms: Option<String>,
    pub container_id: Option<String>,
    pub tag_id: Option<String>,
    pub blocking_rule_id: Option<Vec<String>>,
    pub fingerprint: Option<String>,
    pub firing_rule_id: Option<Vec<String>>,
    pub firing_trigger_id: Option<Vec<String>>,
    pub account_id: Option<String>,
    pub name: Option<String>,
    pub parameter: Option<Vec<Parameter>>,
    pub notes: Option<String>,
    pub priority: Option<Parameter>,
    pub live_only: Option<bool>,
    pub type_: Option<String>,
    pub blocking_trigger_id: Option<Vec<String>>,
}

Represents a Google Tag Manager Tag.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers tags update accounts (request|response)

containers tags get accounts (response)

containers tags create accounts (request|response)

Fields

schedule_start_ms

The start timestamp in milliseconds to schedule a tag.

schedule_end_ms

The end timestamp in milliseconds to schedule a tag.

container_id

GTM Container ID.

tag_id

The Tag ID uniquely identifies the GTM Tag.

blocking_rule_id

Blocking rule IDs. If any of the listed rules evaluate to true, the tag will not fire.

fingerprint

The fingerprint of the GTM Tag as computed at storage time. This value is recomputed whenever the tag is modified.

firing_rule_id

Firing rule IDs. A tag will fire when any of the listed rules are true and all of its blockingRuleIds (if any specified) are false.

firing_trigger_id

Firing trigger IDs. A tag will fire when any of the listed triggers are true and all of its blockingTriggerIds (if any specified) are false.

account_id

GTM Account ID.

name

Tag display name.

parameter

The tag's parameters.

notes

User notes on how to apply this tag in the container.

priority

User defined numeric priority of the tag. Tags are fired asynchronously in order of priority. Tags with higher numeric value fire first. A tag's priority can be a positive or negative value. The default value is 0.

live_only

If set to true, this tag will only fire in the live environment (e.g. not in preview or debug mode).

type_

GTM Tag Type.

blocking_trigger_id

Blocking trigger IDs. If any of the listed triggers evaluate to true, the tag will not fire.