Struct google_tagmanager1::Trigger

pub struct Trigger {
    pub container_id: Option<String>,
    pub unique_trigger_id: Option<Parameter>,
    pub trigger_id: Option<String>,
    pub fingerprint: Option<String>,
    pub enable_all_videos: Option<Parameter>,
    pub account_id: Option<String>,
    pub event_name: Option<Parameter>,
    pub video_percentage_list: Option<Parameter>,
    pub wait_for_tags: Option<Parameter>,
    pub name: Option<String>,
    pub interval: Option<Parameter>,
    pub auto_event_filter: Option<Vec<Condition>>,
    pub filter: Option<Vec<Condition>>,
    pub wait_for_tags_timeout: Option<Parameter>,
    pub limit: Option<Parameter>,
    pub check_validation: Option<Parameter>,
    pub custom_event_filter: Option<Vec<Condition>>,
    pub type_: Option<String>,
    pub parent_folder_id: Option<String>,
}

Represents a Google Tag Manager Trigger

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers triggers update accounts (request|response)

containers triggers get accounts (response)

containers triggers create accounts (request|response)

Fields
container_id: Option<String>

GTM Container ID.

unique_trigger_id: Option<Parameter>

Globally unique id of the trigger that auto-generates this (a Form Submit, Link Click or Timer listener) if any. Used to make incompatible auto-events work together with trigger filtering based on trigger ids. This value is populated during output generation since the tags implied by triggers don't exist until then. Only valid for Form Submit, Link Click and Timer triggers.

trigger_id: Option<String>

The Trigger ID uniquely identifies the GTM Trigger.

fingerprint: Option<String>

The fingerprint of the GTM Trigger as computed at storage time. This value is recomputed whenever the trigger is modified.

enable_all_videos: Option<Parameter>

Reloads the videos in the page that don't already have the YT API enabled. If false, only capture events from videos that already have the API enabled. Only valid for YouTube triggers.

account_id: Option<String>

GTM Account ID.

event_name: Option<Parameter>

Name of the GTM event that is fired. Only valid for Timer triggers.

video_percentage_list: Option<Parameter>

List of integer percentage values. The trigger will fire as each percentage is reached in any instrumented videos. Only valid for YouTube triggers.

wait_for_tags: Option<Parameter>

Whether or not we should delay the form submissions or link opening until all of the tags have fired (by preventing the default action and later simulating the default action). Only valid for Form Submission and Link Click triggers.

name: Option<String>

Trigger display name.

interval: Option<Parameter>

Time between triggering recurring Timer Events (in milliseconds). Only valid for Timer triggers.

auto_event_filter: Option<Vec<Condition>>

Used in the case of auto event tracking.

filter: Option<Vec<Condition>>

The trigger will only fire iff all Conditions are true.

wait_for_tags_timeout: Option<Parameter>

How long to wait (in milliseconds) for tags to fire when 'waits_for_tags' above evaluates to true. Only valid for Form Submission and Link Click triggers.

limit: Option<Parameter>

Limit of the number of GTM events this Timer Trigger will fire. If no limit is set, we will continue to fire GTM events until the user leaves the page. Only valid for Timer triggers.

check_validation: Option<Parameter>

Whether or not we should only fire tags if the form submit or link click event is not cancelled by some other event handler (e.g. because of validation). Only valid for Form Submission and Link Click triggers.

custom_event_filter: Option<Vec<Condition>>

Used in the case of custom event, which is fired iff all Conditions are true.

type_: Option<String>

Defines the data layer event that causes this trigger.

parent_folder_id: Option<String>

Parent folder id.