Struct google_tagmanager1::UserAccess

pub struct UserAccess {
    pub container_access: Option<Vec<ContainerAccess>>,
    pub account_access: Option<AccountAccess>,
    pub email_address: Option<String>,
    pub permission_id: Option<String>,
    pub account_id: Option<String>,
}

Represents a user's permissions to an account and its container.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions get accounts (response)

permissions create accounts (request|response)

permissions update accounts (request|response)

Fields
container_access: Option<Vec<ContainerAccess>>

GTM Container access permissions.

account_access: Option<AccountAccess>

GTM Account access permissions.

email_address: Option<String>

User's email address.

permission_id: Option<String>

Account Permission ID.

account_id: Option<String>

GTM Account ID.