Struct google_tagmanager1::Account

pub struct Account {
    pub share_data: Option<bool>,
    pub account_id: Option<String>,
    pub name: Option<String>,
    pub fingerprint: Option<String>,
}

Represents a Google Tag Manager Account.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers list accounts (none)

permissions list accounts (none)

containers versions undelete accounts (none)

permissions create accounts (none)

permissions delete accounts (none)

containers get accounts (none)

containers versions list accounts (none)

containers triggers update accounts (none)

containers triggers get accounts (none)

containers delete accounts (none)

containers create accounts (none)

containers tags delete accounts (none)

containers rules update accounts (none)

containers rules delete accounts (none)

containers tags list accounts (none)

containers versions publish accounts (none)

containers tags create accounts (none)

containers triggers list accounts (none)

containers versions delete accounts (none)

update accounts (request|response)

containers macros delete accounts (none)

containers versions create accounts (none)

permissions get accounts (none)

containers rules create accounts (none)

containers versions restore accounts (none)

containers rules get accounts (none)

containers variables create accounts (none)

containers variables list accounts (none)

containers macros create accounts (none)

containers tags get accounts (none)

containers variables get accounts (none)

containers triggers delete accounts (none)

containers macros list accounts (none)

containers triggers create accounts (none)

containers macros update accounts (none)

list accounts (none)

permissions update accounts (none)

containers variables delete accounts (none)

get accounts (response)

containers update accounts (none)

containers rules list accounts (none)

containers tags update accounts (none)

containers macros get accounts (none)

containers versions update accounts (none)

containers variables update accounts (none)

containers versions get accounts (none)

Fields

share_data

Whether the account shares data anonymously with Google and others.

account_id

The Account ID uniquely identifies the GTM Account.

name

Account display name.

fingerprint

The fingerprint of the GTM Account as computed at storage time. This value is recomputed whenever the account is modified.