Struct google_tagmanager1::Macro

pub struct Macro {
    pub schedule_start_ms: Option<String>,
    pub schedule_end_ms: Option<String>,
    pub macro_id: Option<String>,
    pub container_id: Option<String>,
    pub parameter: Option<Vec<Parameter>>,
    pub notes: Option<String>,
    pub fingerprint: Option<String>,
    pub disabling_rule_id: Option<Vec<String>>,
    pub enabling_rule_id: Option<Vec<String>>,
    pub account_id: Option<String>,
    pub type_: Option<String>,
    pub parent_folder_id: Option<String>,
    pub name: Option<String>,
}

Represents a Google Tag Manager Macro.

This type is not used in any activity, and only used as part of another schema.

Fields

schedule_start_ms

The start timestamp in milliseconds to schedule a macro.

schedule_end_ms

The end timestamp in milliseconds to schedule a macro.

macro_id

The Macro ID uniquely identifies the GTM Macro.

container_id

GTM Container ID.

parameter

The macro's parameters.

notes

User notes on how to apply this macro in the container.

fingerprint

The fingerprint of the GTM Macro as computed at storage time. This value is recomputed whenever the macro is modified.

disabling_rule_id

For mobile containers only: A list of rule IDs for disabling conditional macros; the macro is enabled if one of the enabling rules is true while all the disabling rules are false. Treated as an unordered set.

enabling_rule_id

For mobile containers only: A list of rule IDs for enabling conditional macros; the macro is enabled if one of the enabling rules is true while all the disabling rules are false. Treated as an unordered set.

account_id

GTM Account ID.

type_

GTM Macro Type.

parent_folder_id

Parent folder id.

name

Macro display name.