Struct google-tagmanager1::Rule

pub struct Rule {
    pub container_id: Option<String>,
    pub name: Option<String>,
    pub rule_id: Option<String>,
    pub notes: Option<String>,
    pub fingerprint: Option<String>,
    pub condition: Vec<Condition>,
    pub account_id: Option<String>,
}

Represents a Google Tag Manager Rule.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers rules get accounts (response)

containers rules create accounts (request|response)

containers rules update accounts (request|response)

Fields

container_id

GTM Container ID.

name

Rule display name.

rule_id

The Rule ID uniquely identifies the GTM Rule.

notes

User notes on how to apply this rule in the container.

fingerprint

The fingerprint of the GTM Rule as computed at storage time. This value is recomputed whenever the rule is modified.

condition

The list of conditions that make up this rule (implicit AND between them).

account_id

GTM Account ID.